\name{depend.truncation-package}
\alias{depend.truncation-package}
\alias{depend.truncation}
\docType{package}
\title{Inference for parametric and nonparametric models with dependent truncation data}
\description{Suppose that one can observe bivariate random variables (X, Y) only when X<=Y holds.
Such data are called truncated data. This package performs parametric and nonparametric 
inference methods for models on (X, Y), especially when X and Y are dependent (correlated)}
\details{
\tabular{ll}{
Package: \tab depend.truncation\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-02-17\cr
License: \tab GPL-2\cr
}
}
\author{
Takeshi EMURA <emura@stat.ncu.edu.tw>
}


