\name{depend.truncation-package}
\alias{depend.truncation-package}
\alias{depend.truncation}
\docType{package}
\title{Statistical Methods for the Analysis of Dependently Truncated Data}
\description{
 Estimation and testing methods for dependently truncated data.
 Semi-parametric methods are based on Emura et al. (2011), Emura & Wang (2012), and Emura & Murotani (2015).
 Parametric approaches are based on Emura & Konno (2012) and Emura & Pan (2017).
 A regression approach is based on Emura & Wang (2016). Quasi-independence tests are based on Emura & Wang (2010).
 Right-truncated data for Japanese male centenarians are given by Emura & Murotani (2015).
}

\details{
\tabular{ll}{
Package: \tab depend.truncation\cr
Type: \tab Package\cr
Version: \tab 3.0\cr
Date: \tab 2018-02-27\cr
License: \tab GPL-2\cr
}
}
\author{
Takeshi Emura <takeshiemura@gmail.com>
}
\references{
Emura T, Wang W (2010), Testing quasi-independence for truncation data. Journal of Multivariate Analysis 101, 223-239.

Emura T, Wang W, Hung HN (2011), Semiparametric Inference for Copula Models for Truncated Data, Stat Sinica 21: 349-367.

Emura T, Konno Y (2012), Multivariate Normal Distribution Approaches for Dependently Truncated Data.
Statistical Papers 53 (No.1), 133-149.

Emura T, Wang W (2012), Nonparametric Maximum Likelihood Estimation for Dependent Truncation Data Based on Copulas, Journal of Multivariate Analysis 110: 171-88

Emura T, Murotani K (2015), An Algorithm for Estimating Survival Under a Copula-based Dependent Truncation Model, TEST 24 (No.4): 734-751.

Emura T, Wang W (2016), Semiparametric Inference for an Accelerated Failure Time Model with Dependent Truncation, Ann Inst Stat Math 68 (5): 1073-94.

Emura T, Pan CH (2017), Parametric likelihood inference and goodness-of-fit for dependently left-truncated data, a copula-based approach, Statistical Papers, doi:10.1007/s00362-017-0947-z.
}