\name{depend.truncation-package}
\alias{depend.truncation-package}
\alias{depend.truncation}
\docType{package}
\title{Inference for parametric and semiparametric models based on dependent truncation data}
\description{
Suppose that one can observe bivariate random variables (X, Y) only when X<=Y holds.
Data (X_j, Y_j), subject X_j<=Y_j, for all j=1,...,n, are called truncated data
 (or truncation data).
 This package performs parametric and semiparametric 
inference methods for models on (X, Y), especially when X and Y are dependent (correlated).
}
\details{
\tabular{ll}{
Package: \tab depend.truncation\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2012-04-2\cr
License: \tab GPL-2\cr
}
}
\author{
Takeshi EMURA <emura@stat.ncu.edu.tw>
}
